refactor: console과 web관련 파일들 분류
