+++ src/main/java/console/controller/Request.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package console.controller;
 
 import chess.domain.board.Point;
