+++ src/main/java/console/view/PieceRepresentation.java
@@ -1,4 +1,4 @@
-package chess.view;
+package console.view;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
